# client/openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"Mcart api documentation","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/":{"get":{"tags":["Home"],"operationId":"health_check","responses":{"200":{"description":"health check","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/auth/forgot-password":{"post":{"tags":["Auth"],"summary":"Forgot Password","operationId":"forgot_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthForgotPasswordInput"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","description":"Api to login for user","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginInput"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/auth/logout":{"delete":{"tags":["Auth"],"summary":"Logout","operationId":"logout","responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterInput"}}},"required":true},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"409":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/auth/register/verify":{"post":{"tags":["Auth"],"summary":"Register Verification","operationId":"register_verify","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterVerifyInput"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"409":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/auth/reset-password":{"post":{"tags":["Auth"],"summary":"Reset Password","operationId":"reset_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResetPasswordInput"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/user/addresses":{"get":{"tags":["Address"],"summary":"Get All Addresses","operationId":"get_addresses","responses":{"200":{"description":"Addresses","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Model"}}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}},"post":{"tags":["Address"],"summary":"Create An Address","operationId":"create_address","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressInput"}}},"required":true},"responses":{"201":{"description":"Address Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressResponse"}}}},"429":{"description":"Address Limit Reached","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/user/addresses/{id}":{"get":{"tags":["Address"],"summary":"Get An Address","operationId":"get_address","parameters":[{"name":"id","in":"path","description":"Address Id","required":true,"schema":{"type":"string","maxLength":24,"minLength":24}}],"responses":{"200":{"description":"Address","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAddressResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}},"delete":{"tags":["Address"],"summary":"Delete An Address","operationId":"delete_address","parameters":[{"name":"id","in":"path","description":"Address Id","required":true,"schema":{"type":"string","maxLength":24,"minLength":24}}],"responses":{"200":{"description":"Address Deleted","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"404":{"description":"Address Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}},"patch":{"tags":["Address"],"summary":"Update An Address","operationId":"update_address","parameters":[{"name":"id","in":"path","description":"Address Id","required":true,"schema":{"type":"string","maxLength":24,"minLength":24}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressInput"}}},"required":true},"responses":{"200":{"description":"Address Updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressResponse"}}}},"404":{"description":"Address Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/user/addresses/{id}/default":{"patch":{"tags":["Address"],"summary":"Mark Default Address","operationId":"mark_default_address","parameters":[{"name":"id","in":"path","description":"Address Id","required":true,"schema":{"type":"string","maxLength":24,"minLength":24}}],"responses":{"200":{"description":"Address Marked as Default","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAddressResponse"}}}},"404":{"description":"Address Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseWithMessage"}}}}}}},"/user/orders":{"get":{"tags":["Order"],"summary":"Get All Orders","operationId":"get_orders","responses":{}}},"/user/orders/id":{"get":{"tags":["Order"],"summary":"Get An Order","operationId":"get_order","responses":{}}},"/user/profile":{"get":{"tags":["Profile"],"summary":"Get Profile","operationId":"get_profile","responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}}}},"patch":{"tags":["Profile"],"summary":"Update Profile","operationId":"update_profile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfile"}}},"required":true},"responses":{"200":{"description":"","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/user/profile/logout-all":{"patch":{"tags":["Profile"],"summary":"Logout All","operationId":"logout_all","responses":{}}},"/user/profile/password":{"patch":{"tags":["Profile"],"summary":"Change Password","operationId":"change_password","responses":{}}}},"components":{"schemas":{"AdminBadRequest":{"type":"object","required":["message3","myName"],"properties":{"message3":{"type":"string"},"myName":{"type":"string"}}},"AdminInternalServerError":{"type":"object","required":["message2","hisName"],"properties":{"hisName":{"type":"string"},"message2":{"type":"string"}}},"AdminLoginInput":{"type":"object","required":["username","password","fullName"],"properties":{"fullName":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}}},"AdminLoginResponse":{"type":"object","required":["message","username","yourName"],"properties":{"message":{"type":"string"},"username":{"type":"string"},"yourName":{"type":"string"}}},"AuthForgotPasswordInput":{"type":"object","required":["email"],"properties":{"email":{"type":"string","description":"Email address of the user.","example":"princeraj9137@gmail.com"}}},"AuthLoginInput":{"type":"object","description":"Credentials of the user.","required":["email","password"],"properties":{"email":{"type":"string","description":"Email address of the user.","example":"princeraj9137@gmail.com"},"password":{"type":"string","description":"Password of the user.","example":"admin790"}}},"AuthRegisterInput":{"type":"object","required":["fullName","email","password","confirmPassword"],"properties":{"confirmPassword":{"type":"string","description":"Password Confirmation.","example":"admin790"},"email":{"type":"string","description":"Email address of the user.","example":"princeraj9137@gmail.com"},"fullName":{"type":"string","description":"Full name of the user.","example":"Jamsr World"},"password":{"type":"string","description":"Password of the user.\nMin. 8 characters.","example":"admin790"}}},"AuthRegisterVerifyInput":{"allOf":[{"$ref":"#/components/schemas/AuthRegisterInput","description":"The original Register struct fields."},{"type":"object","required":["code"],"properties":{"code":{"type":"integer","format":"int32","description":"Verification code (OTP).","example":1234,"minimum":0}}}]},"AuthResetPasswordInput":{"type":"object","required":["email","otp","password","confirmPassword"],"properties":{"confirmPassword":{"type":"string","description":"Password Confirmation.","example":"admin790"},"email":{"type":"string","description":"Email address of the user.","example":"princeraj9137@gmail.com"},"otp":{"type":"integer","format":"int32","description":"Verification code (OTP).","example":1234,"minimum":0},"password":{"type":"string","description":"New Password to set.","example":"admin790"}}},"CreateAddressInput":{"type":"object","required":["firstName","lastName","phoneNumber","fullAddress","city","state","postalCode"],"properties":{"city":{"type":"string","description":"City of the user.","maxLength":50,"minLength":1},"firstName":{"type":"string","description":"First name of the user.","example":"John","maxLength":20,"minLength":1},"fullAddress":{"type":"string","description":"Address of the user.","maxLength":300,"minLength":1},"landmark":{"type":["string","null"],"description":"Landmark of the user.","maxLength":200,"minLength":1},"lastName":{"type":"string","description":"Last name of the user.","example":"Doe","maxLength":20,"minLength":1},"phoneNumber":{"type":"string","description":"Phone number of the user.","example":"1234567890","maxLength":10,"minLength":1},"postalCode":{"type":"integer","format":"int32","description":"Postal Code of the user.","example":123456,"maximum":999999,"minimum":100000},"state":{"type":"string","description":"State of the user.","maxLength":50,"minLength":1}}},"CreateAddressResponse":{"type":"object","required":["message","data"],"properties":{"data":{"$ref":"#/components/schemas/Model"},"message":{"type":"string"}}},"GetAddressResponse":{"type":"object","required":["data"],"properties":{"data":{"$ref":"#/components/schemas/Model"}}},"Model":{"type":"object","required":["id","userId","firstName","lastName","postalCode","city","state","fullAddress","phoneNumber","isDefault","createdAt"],"properties":{"city":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"firstName":{"type":"string"},"fullAddress":{"type":"string"},"id":{"type":"string"},"isDefault":{"type":"boolean"},"landmark":{"type":["string","null"]},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"postalCode":{"type":"integer","format":"int32"},"state":{"type":"string"},"userId":{"type":"string"}}},"ResponseWithMessage":{"type":"object","required":["message"],"properties":{"message":{"type":"string"}}},"UpdateProfile":{"type":"object","required":["full_name"],"properties":{"full_name":{"type":"string"}}},"UserProfile":{"type":"object","required":["id","email","fullName","role"],"properties":{"email":{"type":"string"},"fullName":{"type":"string"},"id":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"}}},"UserRole":{"type":"string","enum":["Admin","User"]}},"responses":{"CreateAddressInput":{"description":"","content":{"application/json":{"schema":{"type":"object","required":["firstName","lastName","phoneNumber","fullAddress","city","state","postalCode"],"properties":{"city":{"type":"string","description":"City of the user.","maxLength":50,"minLength":1},"firstName":{"type":"string","description":"First name of the user.","example":"John","maxLength":20,"minLength":1},"fullAddress":{"type":"string","description":"Address of the user.","maxLength":300,"minLength":1},"landmark":{"type":["string","null"],"description":"Landmark of the user.","maxLength":200,"minLength":1},"lastName":{"type":"string","description":"Last name of the user.","example":"Doe","maxLength":20,"minLength":1},"phoneNumber":{"type":"string","description":"Phone number of the user.","example":"1234567890","maxLength":10,"minLength":1},"postalCode":{"type":"integer","format":"int32","description":"Postal Code of the user.","example":123456,"maximum":999999,"minimum":100000},"state":{"type":"string","description":"State of the user.","maxLength":50,"minLength":1}}}}}}}}}